pyproject.toml
[project]
name = "spotdl-cask"
version = "0.1.0"
description = "spotDL Web Downloader desktop app"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "flask==3.1.3",
    "pywebview==6.1",
    "spotdl==4.4.3",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.0",
    "ruff",
    "ty>=0.0.26",
]

[tool.ruff]
target-version = "py311"
extend-exclude = [
    "build",
    "dist",
]

[tool.ruff.lint]
extend-select = ["I"]
